(refactor) 할인값 double로 수정 & 중괄호 생략 되돌리기 & 제한값 상수화 & [VoucherProvider] 예외메세지 수정

+++ src/main/java/org/prgrms/kdt/voucher/Voucher.java
@@ -4,9 +4,9 @@ import java.util.UUID;
 
 public interface Voucher {
 
-    long discount(long beforeDiscount);
+    double discount(double beforeDiscount);
 
     UUID getVoucherId();
 
-    long getAmount();
+    double getAmount();
 }
